requirements.txt
Django==2.2.1
django-admin-sortable2==0.7.1
djangorestframework==3.9.4
django-filter==2.1.0
drf-writable-nested==0.5.1
django-behaviors==0.4.1
django-cors-headers==3.0.1
django-tinymce==2.8.0
django-wkhtmltopdf==3.2.0
qrcodegen==1.4.0
Pillow==6.0.0
slackclient==2.0.1